+++ Controllers/SiteController.cs
@@ -4,6 +4,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using MyResourcesApp.Models;
+using System.Net.Mail;
+using System.Net;
+using Enums;
 //using Amazon.DynamoDBv2;
 
 namespace MyResourcesApp.Controllers
@@ -33,17 +36,8 @@ namespace MyResourcesApp.Controllers
         {
             if (ModelState.IsValid)
             {
-                //var dbEntry = _db.siteInfo.FirstOrDefault(acc => acc.CustomerID == site.CustomerID);
-
-                var myUser = _db.site.SingleOrDefault(user => user.SiteName == site.SiteName && user.CustomerID == site.CustomerID);
-                //Func<Site, bool> expression = g => g.CustomerID == site.CustomerID &&
-                //                              g.SiteName == site.SiteName;
-                //var myUser = _db.siteInfo.SingleOrDefault(expression);
-
-                //geoLocation.FirstOrDefault(g => g.Longitude != null && g.Latitude != null);
-
-
-                if (myUser?.SiteName == site.SiteName)
+                var siteDetails = _db.site.SingleOrDefault(s => s.SiteName == site.SiteName && s.CustomerID == site.CustomerID);
+                if (siteDetails?.SiteName == site.SiteName)
                 {
                     ViewBag.CustomerID = site.CustomerID;
                     ViewBag.SiteName = site.SiteName;
@@ -67,6 +61,13 @@ namespace MyResourcesApp.Controllers
                 return RedirectToAction("RegisterSite");
             }
             var getSiteDetails = await _db.site.FindAsync(siteId);
+           var getOrderDetails = _db.order.FirstOrDefault(o => o.SiteName == getSiteDetails.SiteName && o.CID == getSiteDetails.CustomerID && o.OrderStatusID == (char)OrderStatus.Pending);
+            if (getOrderDetails != null)
+            {
+                ViewBag.CustomerID = getSiteDetails.CustomerID;
+                ViewBag.SiteName = getSiteDetails.SiteName;
+                return View("PendingOrder_Exists");
+            }
             return View(getSiteDetails);
         }
 
@@ -97,12 +98,23 @@ namespace MyResourcesApp.Controllers
         [HttpPost]
         public async Task<IActionResult> DeleteSiteInfo(int siteId)
         {
-
             var getSiteDetails = await _db.site.FindAsync(siteId);
-            _db.site.Remove(getSiteDetails);
-            await _db.SaveChangesAsync();
-            TempData["message"] = siteId + " was deleted";
-            return RedirectToAction("RegisterSite");
+            var getOrderDetails = _db.order.FirstOrDefault(o => o.SiteName == getSiteDetails.SiteName && o.CID == getSiteDetails.CustomerID && o.OrderStatusID == (char)OrderStatus.Pending);
+            if (getOrderDetails != null)
+            {
+                ViewBag.CustomerID = getSiteDetails.CustomerID;
+                ViewBag.SiteName = getSiteDetails.SiteName;
+                return View("PendingOrder_Exists");
+            }
+            else
+            {
+                _db.site.Remove(getSiteDetails);
+                await _db.SaveChangesAsync();
+ 
+                return RedirectToAction("RegisterSite");
+            }
+
+            //return RedirectToAction("RegisterSite");
         }
     }
 }
